+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.cornellappdev.android.volume"
+        applicationId "com.appdev.volume_android"
         minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
