+++ app/src/main/java/com/appdev/volume_android/adapters/CustomPagerAdapter.kt
@@ -1,11 +1,11 @@
-package com.cornellappdev.android.volume.adapters
+package com.appdev.volume_android.adapters
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
-import com.cornellappdev.android.volume.fragments.HomeFragment
-import com.cornellappdev.android.volume.fragments.PublicationsFragment
-import com.cornellappdev.android.volume.fragments.SavedPublicationsFragment
+import com.appdev.volume_android.fragments.HomeFragment
+import com.appdev.volume_android.fragments.PublicationsFragment
+import com.appdev.volume_android.fragments.SavedPublicationsFragment
 
 class CustomPagerAdapter(fragmentManager: FragmentManager, private val numOfTabs: Int)
     : FragmentPagerAdapter(fragmentManager, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
