+++ app/src/main/java/com/appdev/volume_android/adapters/FollowPublicationsAdapter.kt
@@ -1,4 +1,4 @@
-package com.cornellappdev.android.volume.adapters
+package com.appdev.volume_android.adapters
 
 import PrefUtils
 import android.content.Context
@@ -12,9 +12,9 @@ import android.widget.TextView
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import com.cornellappdev.android.volume.PublicationProfileActivity
-import com.cornellappdev.android.volume.R
-import com.cornellappdev.android.volume.models.Publication
+import com.appdev.volume_android.PublicationProfileActivity
+import com.appdev.volume_android.R
+import com.appdev.volume_android.models.Publication
 import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.publication_card.view.*
 import kotlinx.android.synthetic.main.publication_profile_activity.*
