+++ app/src/main/java/com/appdev/volume_android/adapters/SavedArticlesAdapter.kt
@@ -1,4 +1,4 @@
-package com.cornellappdev.android.volume.adapters
+package com.appdev.volume_android.adapters
 
 import android.content.Intent
 import android.os.Build
@@ -9,47 +9,49 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.annotation.RequiresApi
-import androidx.cardview.widget.CardView
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.RecyclerView
-import com.cornellappdev.android.volume.MainActivity
-import com.cornellappdev.android.volume.R
-import com.cornellappdev.android.volume.models.Article
+import com.appdev.volume_android.MainActivity
+import com.appdev.volume_android.R
+import com.appdev.volume_android.models.Article
 import com.squareup.picasso.Picasso
-import kotlinx.android.synthetic.main.article_card.view.*
+import kotlinx.android.synthetic.main.saved_article.view.*
+import kotlinx.android.synthetic.main.saved_article.view.article_img_b
+import kotlinx.android.synthetic.main.saved_article.view.article_shout_count_b
+import kotlinx.android.synthetic.main.saved_article.view.article_title_b
+import kotlinx.android.synthetic.main.saved_article.view.desc_holder_b
+import kotlinx.android.synthetic.main.saved_article.view.post_time_b
+import kotlinx.android.synthetic.main.saved_article.view.pub_name_article_layout_b
+import kotlinx.android.synthetic.main.saved_article.view.saved_article_layout
+import kotlinx.android.synthetic.main.saved_article.view.*
 import java.time.Duration
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
-class ArticleAdapter(private val articles: List<Article>) :
-        RecyclerView.Adapter<ArticleAdapter.ArticleVH>() {
+class SavedArticlesAdapter(private val articles: List<Article>) :
+        RecyclerView.Adapter<SavedArticlesAdapter.SavedArticleVH>() {
 
-    class ArticleVH(itemView : View) : RecyclerView.ViewHolder(itemView){
-
-        val articleTitle : TextView = itemView.article_title
-        val articleImg : ImageView = itemView.article_img
-        val postTime: TextView = itemView.post_time
-        val shoutoutCount: TextView = itemView.article_shout_count
-
-        val layout: ConstraintLayout = itemView.article_layout
-
+    class SavedArticleVH(itemView : View) : RecyclerView.ViewHolder(itemView) {
+        val pubName: TextView = itemView.pub_name_article_layout_b
+        val articleTitle : TextView = itemView.article_title_b
+        val articleImg : ImageView = itemView.article_img_b
+        val postTime: TextView = itemView.post_time_b
+        val shoutoutCount: TextView = itemView.article_shout_count_b
+        val layout: ConstraintLayout = itemView.desc_holder_b
+        val layoutMain: ConstraintLayout = itemView.saved_article_layout
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ArticleVH {
-        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.article_card, parent, false)
-        return ArticleVH(itemView)
-
-    }
-
-    override fun getItemCount(): Int {
-        return articles.size
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int):  SavedArticleVH {
+        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.saved_article, parent, false)
+        return SavedArticleVH (itemView)
     }
 
     @RequiresApi(Build.VERSION_CODES.O)
-    override fun onBindViewHolder(holder: ArticleVH, position: Int) {
+    override fun onBindViewHolder(holder: SavedArticleVH, position: Int) {
         val currentItem : Article = articles[position]
         holder.articleTitle.text = currentItem.title
-        if(currentItem.imageURL != null && currentItem.imageURL != ""){
+        if(!currentItem.imageURL.isNullOrBlank()) {
+            holder.articleImg.visibility = View.VISIBLE
             Picasso.get().load(currentItem.imageURL).into(holder.articleImg)
         }
         val format = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
@@ -91,10 +93,16 @@ class ArticleAdapter(private val articles: List<Article>) :
             }
         }
         holder.shoutoutCount.text = currentItem.shoutouts?.toInt().toString() + " shout-outs"
-        holder.layout.setOnClickListener{
+        holder.pubName.text = currentItem.publication?.name
+
+        holder.layoutMain.setOnClickListener{
             val intent = Intent(holder.layout.context, MainActivity::class.java)
             intent.putExtra("article",currentItem)
             holder.layout.context?.startActivity(intent)
         }
+    }
+
+    override fun getItemCount(): Int {
+        return articles.size
     }
 }
