+++ app/src/main/java/com/appdev/volume_android/fragments/OnboardingFrag1.kt
@@ -1,11 +1,11 @@
-package com.cornellappdev.android.volume.fragments
+package com.appdev.volume_android.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.cornellappdev.android.volume.R
+import com.appdev.volume_android.R
 
 class OnboardingFrag1 : Fragment() {
